refactor: 서비스 레이어 Critical 4건 + Major 12건 리팩토링
Phase 1: 기반 인프라 생성
- StringMaskingUtils: 7개 파일의 마스킹 중복 제거 (maskOcid, maskFingerprint 등 5개 메서드)
- Custom Exception 4개: CachePersistenceException, UnsupportedCalculationEngineException,
  TransactionSnapshotException, InvalidAdminFingerprintException

Phase 2: Critical 4건 수정
- RuntimeException → CachePersistenceException (EquipmentCacheService)
- 안전하지 않은 캐스트 → instanceof 패턴 매칭 (4개 파일)
- UnsupportedOperationException → UnsupportedCalculationEngineException (CubeServiceImpl)
- IllegalStateException → TransactionSnapshotException (EquipmentService)

Phase 3: Major 서비스 레벨 수정
- 7개 파일의 private maskXxx() → StringMaskingUtils 위임
- AdminService: IllegalArgumentException → InvalidAdminFingerprintException
- CharacterLikeService: 안전한 문자열 서브스트링 (StringIndexOutOfBounds 방지)
- AuthService: 주석 번호 수정 (7→8, 8→9)
- GameCharacterService: @Async self-invocation ObjectProvider 수정

Phase 4: 구조적 리팩토링
- PresetCalculationHelper: V4 God Class 분해 (444줄 → ~250줄)
- LikeSyncMetricsRecorder: 메트릭 관심사 분리 (LikeSyncService 440줄 → ~380줄)
- CharacterCreationService: OcidResolver/GameCharacterService 공통 로직 추출

신규 8개 파일, 수정 16개 파일 (테스트 포함)
